requirements.txt
PyYAML==3.13
pika==0.12.0
tornado==5.1.1
python-dateutil==2.7.5